Add files via upload
This dashboard, was created as part of one of the projects during my MS.B A program, showing the Key performance industry of a store. This interactive dashboard can be used to identify the sales, profit, shipping cost, Quantity, and total transactions which are few of the main KPI's for any sales business. These variables are further compared using the regions, categories and sub-categories.